requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
mysql-connector-python>=8.0.0
cryptography>=41.0.0
bcrypt>=4.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
requests>=2.31.0
gunicorn>=21.0.0